+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @claudexor/gateway
 
+## 3.12.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @claudexor/schema@3.12.10
+  - @claudexor/core@3.12.10
+
 ## 3.12.9
 
 ### Patch Changes