+++ packages/synthesis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @claudexor/synthesis
 
+## 3.12.10
+
+### Patch Changes
+
+- @claudexor/arbitration@3.12.10
+
 ## 3.12.9
 
 ### Patch Changes